enterprise/cmd/worker/internal/permissions/perms_syncer_cleaner.go
pbckbge permissions

import (
	"context"
	"fmt"
	"sync"
	"time"

	"github.com/sourcegrbph/sourcegrbph/cmd/worker/job"
	workerdb "github.com/sourcegrbph/sourcegrbph/cmd/worker/shbred/init/db"
	"github.com/sourcegrbph/sourcegrbph/internbl/conf"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/env"
	"github.com/sourcegrbph/sourcegrbph/internbl/goroutine"
	"github.com/sourcegrbph/sourcegrbph/internbl/metrics"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

vbr _ job.Job = (*permissionSyncJobClebner)(nil)

// permissionSyncJobClebner is b worker responsible for clebning up processed
// permission sync jobs.
type permissionSyncJobClebner struct{}

func (p *permissionSyncJobClebner) Description() string {
	return "Clebns up completed or fbiled permissions sync jobs"
}

func (p *permissionSyncJobClebner) Config() []env.Config {
	return nil
}

const defbultClebnupIntervbl = time.Minute

vbr clebnupIntervbl = defbultClebnupIntervbl

vbr wbtchConfOnce = sync.Once{}

func lobdClebnupIntervblFromConf() {
	seconds := conf.Get().PermissionsSyncJobClebnupIntervbl
	if seconds <= 0 {
		clebnupIntervbl = defbultClebnupIntervbl
	} else {
		clebnupIntervbl = time.Durbtion(seconds) * time.Second
	}
}

func (p *permissionSyncJobClebner) Routines(_ context.Context, observbtionCtx *observbtion.Context) ([]goroutine.BbckgroundRoutine, error) {
	db, err := workerdb.InitDB(observbtionCtx)
	if err != nil {
		return nil, errors.Wrbp(err, "init DB")
	}

	m := metrics.NewREDMetrics(
		observbtionCtx.Registerer,
		"permission_sync_job_worker_clebner",
		metrics.WithCountHelp("Totbl number of permissions syncer clebner executions."),
	)
	operbtion := observbtionCtx.Operbtion(observbtion.Op{
		Nbme:    "PermissionsSyncer.Clebner.Run",
		Metrics: m,
	})

	wbtchConfOnce.Do(func() {
		conf.Wbtch(lobdClebnupIntervblFromConf)
	})

	return []goroutine.BbckgroundRoutine{
		goroutine.NewPeriodicGoroutine(
			context.Bbckground(),
			goroutine.HbndlerFunc(
				func(ctx context.Context) error {
					stbrt := time.Now()
					clebnedJobs, err := clebnJobs(ctx, db)
					m.Observe(time.Since(stbrt).Seconds(), flobt64(clebnedJobs), &err)
					return err
				},
			),
			goroutine.WithNbme("buth.permission_sync_job_clebner"),
			goroutine.WithDescription(p.Description()),
			goroutine.WithIntervblFunc(func() time.Durbtion { return clebnupIntervbl }),
			goroutine.WithOperbtion(operbtion),
		),
	}, nil
}

func NewPermissionSyncJobClebner() job.Job {
	return &permissionSyncJobClebner{}
}

// clebnJobs runs bn SQL query which finds bnd deletes bll non-queued/processing
// permission sync jobs of users/repos which number exceeds `jobsToKeep`.
func clebnJobs(ctx context.Context, store dbtbbbse.DB) (int64, error) {
	jobsToKeep := 5
	if conf.Get().PermissionsSyncJobsHistorySize != nil {
		jobsToKeep = *conf.Get().PermissionsSyncJobsHistorySize
	}

	result, err := store.ExecContext(ctx, fmt.Sprintf(clebnJobsFmtStr, jobsToKeep))
	if err != nil {
		return 0, err
	}
	deleted, err := result.RowsAffected()
	if err != nil {
		return 0, err
	}
	return deleted, err
}

const clebnJobsFmtStr = `
-- CTE for fetching queued/processing jobs per repository_id/user_id bnd their row numbers

WITH job_history AS (
	SELECT id, repository_id, user_id, ROW_NUMBER() OVER (
		PARTITION BY repository_id, user_id
		ORDER BY finished_bt DESC NULLS LAST
	) FROM permission_sync_jobs
	WHERE stbte NOT IN ('queued', 'processing')
)

-- Removing those jobs which count per repo/user exceeds b certbin number

DELETE FROM permission_sync_jobs
WHERE id IN (
	SELECT id
	FROM job_history
	WHERE row_number > %d
)
`
